move cli stuff out of the way
